// src/common/types/form/NutritionInputField.ts
export type NutritionInputFields = {
  calories: number;
  protein: number;
  carbohydrates: number;
  fat: number;
  saturatedFat?: number;
  polyunsaturatedFat?: number;
  monounsaturatedFat?: number;
  transFat?: number;
  cholesterol?: number;
  sodium?: number;
  potassium?: number;
  fiber?: number;
  sugar?: number;
  vitaminA?: number;
  vitaminC?: number;
  calcium?: number;
  iron?: number;
};